% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxSurvey.R
\name{rake_wt}
\alias{rake_wt}
\title{Weighted or Unweighted Raking}
\usage{
rake_wt(
  svysmpl,
  svypopu,
  auxVars,
  svyVar,
  subset = NULL,
  family = gaussian(),
  invlvls,
  weights = NULL,
  maxiter = 50
)
}
\arguments{
\item{svysmpl}{A dataframe or tibble to represent 'samples'.}

\item{svypopu}{A dataframe or tibble to represent 'population'.}

\item{auxVars}{A character vector contains the names of auxiliary variables for raking.}

\item{svyVar}{Outcome variable.}

\item{subset}{A character vector. Each element is a string representing a filtering condition to select subset of samples and population. Default is NULL. When this parameter is NULL, the analysis is only performed on the whole data. If subsets are specified, the estimates for the whole data will also be calculated.}

\item{family}{The distribution family of outcome variable. Currently we only support \code{\link[stats]{gaussian}} and \code{\link[stats]{binomial}}.}

\item{invlvls}{A numeric vector of values. Each specifies a confidence level of CI for estimators. If more than one values are specified, then multiple CIs are calculated.}

\item{weights}{A numeric vector of case weights. The length should be equal to the number of cases in 'samples'.}

\item{maxiter}{A integer to specify maximum iteration of raking.}
}
\value{
A list. Each element contains the raking estimate and CIs for a subset or the whole data analysis.
}
\description{
Weighted or Unweighted Raking
}
