% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxSurvey.R
\name{simulate}
\alias{simulate}
\title{Simulate}
\usage{
simulate(N = 3000, discretize = c(3, 5, 10), setting = c(1, 2, 3), seed = NULL)
}
\arguments{
\item{N}{Number of population size}

\item{discretize}{Number of discretized categories}

\item{setting}{Simulation setting}

\item{seed}{Random seed}
}
\value{
simulated dataset
}
\description{
Simulate
}
